requirements.txt
pyzbar
qrcode[pil]
python-barcode
pillow
colorama
